refactor: reorganize Profile structure and move sidebars to section
Major restructuring to fix naming inconsistencies and improve component organization:

## 1. Profile Structure Redesign

Clarified that Profile encompasses both Portfolio and Team:

```
ui/profile/                  # Profile = Portfolio + Team
  ├── ProfileImage.tsx       # Common (used by both)
  ├── ProfileIcon.tsx        # Common (used by both)
  ├── ProfileItem.tsx        # Common (used by both)
  ├── portfolio/             # Portfolio-specific components
  │   ├── ProfileEditButton.tsx
  │   ├── ProfileEditModal.tsx
  │   ├── ProfileInfo.tsx
  │   ├── PortfolioDetail.tsx
  │   ├── PortfolioProject.tsx
  │   └── ProjectImages.tsx
  └── team/                  # Team-specific components
      └── TeamProfileEditButton.tsx
```

File moves:
- feature/portfolio/ProfileItem.tsx → ui/profile/ProfileItem.tsx
- card/portfolio/ProfileEdit*.tsx → ui/profile/portfolio/
- card/portfolio/components/Profile*.tsx → ui/profile/portfolio/
- card/portfolio/components/Portfolio*.tsx → ui/profile/portfolio/
- card/portfolio/components/ProjectImages.tsx → ui/profile/portfolio/
- card/portfolio/TeamProfileEditButton.tsx → ui/profile/team/

## 2. Sidebars Moved to Section

Moved domain-specific sidebars from feature/ to section/ (as they define page sections):

```
section/
  ├── project/
  │   ├── ProjectSidebar.tsx      ← from feature/project/
  │   ├── ProjectSummarySection.tsx
  │   └── ...
  └── team/                        ← newly created
      └── TeamSidebar.tsx          ← from feature/team/
```

## Import Path Updates

Updated imports in 10+ files:
- ProfileItem: TeamSidebar, PortfolioDetail
- Portfolio components: PortfolioHome, PortfolioCard, Portfolio.tsx, etc.
- Team components: TeamSidebar
- Sidebars: ProjectPage.tsx, Team.tsx

Fixed relative imports to absolute paths:
- PortfolioProject.tsx: Card import
- ProfileEditModal.tsx: InputOfModal import
- ProjectImages.tsx: types import

## Benefits

- Clear hierarchy: Profile > {Portfolio, Team}
- Better component discoverability
- Consistent classification: sidebars are sections
- All imports use absolute paths

Verified with pnpm lint and pnpm build.
