+++ src/app/components/card/portfolio/PortfolioHome.tsx
@@ -1,9 +1,9 @@
-import PortfolioItems from './components/PortfolioDetail';
+import PortfolioItems from '@/app/components/ui/profile/portfolio/PortfolioDetail';
 import { Title } from '../../ui/text/text';
 
 import { PortfolioDetailProps } from '@/app/(box-layout)/portfolio/types';
 import { CardProps } from '@/app/components/card/project/ProjectCard';
-import ProfileEditButton from './ProfileEditButton';
+import ProfileEditButton from '@/app/components/ui/profile/portfolio/ProfileEditButton';
 import ProjectGrid from '@/app/components/feature/project/components/ProjectGrid';
 
 interface PortfolioHomeProps {
